+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 sbt-play-messages Changelog
 ===========================
 
-v1.0.0 | 2014-06-09
+v1.0.0-RC2 | 2014-06-??
+
+ * Added a new requireDefaultMessagesFile setting for projects that do not have a default messages file.
+ * Implemented the checkApplicationLanguages functionality.
+
+v1.0.0-RC1 | 2014-06-18
 
  * Initial release with support for Play framework 2.3, Scala and Java.
-